Split up detail-definitions.js
