+++ webofneeds/won-owner-webapp/src/main/webapp/config/detail-definitions.js
@@ -1,12 +1,8 @@
-import {
-  get,
-  getIn,
-  isValidDate,
-  parseDatetimeStrictly,
-  toLocalISODateString,
-  generateIdString,
-} from "../app/utils.js";
-import Immutable from "immutable";
+import personDetails from "./details/person.js";
+import locationDetails from "./details/location.js";
+import timeDetails from "./details/datetime.js";
+import fileDetails from "./details/files.js";
+import priceDetails from "./details/price.js";
 import won from "../app/won-es6.js";
 
 export const emptyDraft = {
@@ -138,7 +134,7 @@ export const abstractDetails = {
   },
 };
 
-export const details = {
+const basicDetails = {
   title: {
     identifier: "title",
     label: "Title",
@@ -159,108 +155,6 @@ export const details = {
     generateHumanReadable: function({ value, includeLabel }) {
       if (value) {
         return includeLabel ? this.label + ": " + value : value;
-      }
-      return undefined;
-    },
-  },
-  price: {
-    identifier: "price",
-    label: "Price",
-    icon: "#ico36_detail_price",
-    placeholder: "Enter Price...",
-    component: "won-price-picker",
-    viewerComponent: "won-price-viewer",
-    currency: [
-      { value: "EUR", label: "€", default: true },
-      { value: "USD", label: "$" },
-      { value: "GBP", label: "£" },
-    ],
-    unitCode: [
-      { value: "MON", label: "per month" },
-      { value: "WEE", label: "per week" },
-      { value: "DAY", label: "per day" },
-      { value: "HUR", label: "per hour" },
-      { value: "", label: "total", default: true },
-    ],
-    messageEnabled: true,
-    parseToRDF: function({ value, identifier, contentUri }) {
-      if (!value || !value.amount || !value.currency) {
-        return { "s:priceSpecification": undefined };
-      }
-
-      return {
-        "s:priceSpecification": {
-          "@id":
-            contentUri && identifier
-              ? contentUri + "/" + identifier + "/" + generateIdString(10)
-              : undefined,
-          "@type": "s:CompoundPriceSpecification",
-          "s:price": [{ "@value": value.amount, "@type": "xsd:float" }],
-          "s:priceCurrency": value.currency,
-          "s:unitCode": value.unitCode,
-        },
-      };
-    },
-    parseFromRDF: function(jsonLDImm) {
-      const amount = won.parseFrom(
-        jsonLDImm,
-        ["s:priceSpecification", "s:price"],
-        "xsd:float"
-      );
-
-      const currency = won.parseFrom(
-        jsonLDImm,
-        ["s:priceSpecification", "s:priceCurrency"],
-        "xsd:string"
-      );
-
-      const unitCode = won.parseFrom(
-        jsonLDImm,
-        ["s:priceSpecification", "s:unitCode"],
-        "xsd:string"
-      );
-
-      if (!amount || !currency) {
-        return undefined;
-      } else {
-        return { amount: amount, currency: currency, unitCode: unitCode };
-      }
-    },
-    generateHumanReadable: function({ value, includeLabel }) {
-      if (value) {
-        const amount = value.amount;
-
-        let currencyLabel = undefined;
-        let unitCodeLabel = undefined;
-
-        this.currency &&
-          this.currency.forEach(curr => {
-            if (curr.value === value.currency) {
-              currencyLabel = curr.label;
-            }
-          });
-        currencyLabel = currencyLabel || value.currency;
-
-        this.unitCode &&
-          this.unitCode.forEach(uc => {
-            if (uc.value === value.unitCode) {
-              unitCodeLabel = uc.label;
-            }
-          });
-        unitCodeLabel = unitCodeLabel || value.unitCode;
-
-        if (unitCodeLabel) {
-          return (
-            (includeLabel ? this.label + ": " + amount : amount) +
-            currencyLabel +
-            " " +
-            unitCodeLabel
-          );
-        } else {
-          return (
-            (includeLabel ? this.label + ": " + amount : amount) + currencyLabel
-          );
-        }
       }
       return undefined;
     },
@@ -292,653 +186,6 @@ export const details = {
     generateHumanReadable: function({ value, includeLabel }) {
       if (value) {
         return includeLabel ? this.label + ": " + value : value;
-      }
-      return undefined;
-    },
-  },
-  fromDatetime: {
-    identifier: "fromDatetime",
-    label: "Starting at",
-    icon: "#ico36_detail_datetime",
-    placeholder: "Enter Date and Time...",
-    component: "won-datetime-picker",
-    viewerComponent: "won-datetime-viewer",
-    messageEnabled: true,
-    parseToRDF: function({ value }) {
-      // value can be an xsd:datetime-string or a javascript date object
-      const datetime = parseDatetimeStrictly(value);
-      if (!isValidDate(datetime)) {
-        return { "s:validFrom": undefined };
-      } else {
-        const datetimeString = toLocalISODateString(datetime);
-        return {
-          "s:validFrom": { "@value": datetimeString, "@type": "xsd:dateTime" },
-        };
-      }
-    },
-    parseFromRDF: function(jsonLDImm) {
-      return won.parseFrom(jsonLDImm, ["s:validFrom"], "xsd:dateTime");
-    },
-    generateHumanReadable: function({ value, includeLabel }) {
-      if (value) {
-        const maybeLabel = includeLabel ? this.label + ": " : "";
-        const datetime = parseDatetimeStrictly(value);
-        const timestring = isValidDate(datetime)
-          ? datetime.toLocaleString()
-          : "";
-        return maybeLabel + timestring;
-      }
-      return undefined;
-    },
-  },
-  throughDatetime: {
-    identifier: "throughDatetime",
-    label: "Ending at",
-    icon: "#ico36_detail_datetime",
-    placeholder: "Enter Date and Time...",
-    component: "won-datetime-picker",
-    viewerComponent: "won-datetime-viewer",
-    messageEnabled: true,
-    parseToRDF: function({ value }) {
-      // value can be an xsd:datetime-string or a javascript date object
-      const datetime = parseDatetimeStrictly(value);
-      if (!isValidDate(datetime)) {
-        return { "s:validThrough": undefined };
-      } else {
-        const datetimeString = toLocalISODateString(datetime);
-        return {
-          "s:validThrough": {
-            "@value": datetimeString,
-            "@type": "xsd:dateTime",
-          },
-        };
-      }
-    },
-    parseFromRDF: function(jsonLDImm) {
-      return won.parseFrom(jsonLDImm, ["s:validThrough"], "xsd:dateTime");
-    },
-    generateHumanReadable: function({ value, includeLabel }) {
-      if (value) {
-        const maybeLabel = includeLabel ? this.label + ": " : "";
-        const datetime = parseDatetimeStrictly(value);
-        const timestring = isValidDate(datetime)
-          ? datetime.toLocaleString()
-          : "";
-        return maybeLabel + timestring;
-      }
-      return undefined;
-    },
-  },
-  datetimeRange: {
-    identifier: "datetimeRange",
-    label: "Timerange",
-    icon: "#ico36_detail_datetime",
-    placeholder: undefined,
-    component: "won-datetime-range-picker",
-    viewerComponent: "won-datetime-viewer", // not used, values are read as fromDatetime and throughDatetime
-    messageEnabled: false,
-    parseToRDF: function({ value }) {
-      if (value) {
-        const fromDatetime = parseDatetimeStrictly(value.fromDatetime);
-        const toDatetime = parseDatetimeStrictly(value.toDatetime);
-
-        let validFrom = undefined;
-        let validThrough = undefined;
-
-        if (isValidDate(fromDatetime)) {
-          const fromDatetimeString = toLocalISODateString(fromDatetime);
-          validFrom = {
-            "@value": fromDatetimeString,
-            "@type": "xsd:dateTime",
-          };
-        }
-
-        if (isValidDate(toDatetime)) {
-          const toDatetimeString = toLocalISODateString(toDatetime);
-          validThrough = {
-            "@value": toDatetimeString,
-            "@type": "xsd:dateTime",
-          };
-        }
-
-        return {
-          "s:validFrom": validFrom,
-          "s:validThrough": validThrough,
-        };
-      } else {
-        return {
-          "s:validFrom": undefined,
-          "s:validThrough": undefined,
-        };
-      }
-    },
-    parseFromRDF: function() {}, // RDF should be parsed by fromDatetime and throughDatetime
-    generateHumanReadable: function() {}, // should be generated by fromDatetime and troughDatetime
-  },
-  location: {
-    identifier: "location",
-    label: "Location",
-    icon: "#ico36_detail_location",
-    placeholder: "Search for location",
-    component: "won-location-picker",
-    viewerComponent: "won-location-viewer",
-    messageEnabled: true,
-    parseToRDF: function({ value, identifier, contentUri }) {
-      if (!value) {
-        return { "won:hasLocation": undefined };
-      }
-
-      const randomLocationId = generateIdString(10);
-
-      return {
-        "won:hasLocation": {
-          "@id":
-            contentUri && identifier
-              ? contentUri + "/" + identifier + "/" + randomLocationId
-              : undefined,
-          "@type": "s:Place",
-          "s:geo": {
-            "@id":
-              contentUri && identifier
-                ? contentUri +
-                  "/" +
-                  identifier +
-                  "/" +
-                  randomLocationId +
-                  "/locationgeo"
-                : undefined,
-            "@type": "s:GeoCoordinates",
-            "s:latitude": value.lat.toFixed(6),
-            "s:longitude": value.lng.toFixed(6),
-            "won:geoSpatial": {
-              "@type":
-                "http://www.bigdata.com/rdf/geospatial/literals/v1#lat-lon",
-              "@value": `${value.lat.toFixed(6)}#${value.lng.toFixed(6)}`,
-            },
-          },
-          "s:name": value.name,
-          "won:hasBoundingBox":
-            !value.nwCorner || !value.seCorner
-              ? undefined
-              : {
-                  "@id":
-                    contentUri && identifier
-                      ? contentUri +
-                        "/" +
-                        identifier +
-                        "/" +
-                        randomLocationId +
-                        "/Bounds"
-                      : undefined,
-                  "won:hasNorthWestCorner": {
-                    "@id":
-                      contentUri && identifier
-                        ? contentUri +
-                          "/" +
-                          identifier +
-                          "/" +
-                          randomLocationId +
-                          "/Bounds/NW"
-                        : undefined,
-                    "@type": "s:GeoCoordinates",
-                    "s:latitude": value.nwCorner.lat.toFixed(6),
-                    "s:longitude": value.nwCorner.lng.toFixed(6),
-                  },
-                  "won:hasSouthEastCorner": {
-                    "@id":
-                      contentUri && identifier
-                        ? contentUri +
-                          "/" +
-                          identifier +
-                          "/" +
-                          randomLocationId +
-                          "/Bounds/SE"
-                        : undefined,
-                    "@type": "s:GeoCoordinates",
-                    "s:latitude": value.seCorner.lat.toFixed(6),
-                    "s:longitude": value.seCorner.lng.toFixed(6),
-                  },
-                },
-        },
-      };
-    },
-    parseFromRDF: function(jsonLDImm) {
-      const jsonldLocation = jsonLDImm && jsonLDImm.get("won:hasLocation");
-      if (!jsonldLocation) return undefined; // NO LOCATION PRESENT
-
-      const jsonldLocationImm = Immutable.fromJS(jsonldLocation);
-
-      let location = {
-        address: undefined,
-        lat: undefined,
-        lng: undefined,
-        nwCorner: {
-          lat: undefined,
-          lng: undefined,
-        },
-        seCorner: {
-          lat: undefined,
-          lng: undefined,
-        },
-      };
-
-      location.address = won.parseFrom(
-        jsonldLocationImm,
-        ["s:name"],
-        "xsd:string"
-      );
-
-      const parseFloatFromLocation = path =>
-        won.parseFrom(jsonldLocationImm, path, "xsd:float");
-
-      location.lat = parseFloatFromLocation(["s:geo", "s:latitude"]);
-      location.lng = parseFloatFromLocation(["s:geo", "s:longitude"]);
-      location.nwCorner.lat = parseFloatFromLocation([
-        "won:hasBoundingBox",
-        "won:hasNorthWestCorner",
-        "s:latitude",
-      ]);
-      location.nwCorner.lng = parseFloatFromLocation([
-        "won:hasBoundingBox",
-        "won:hasNorthWestCorner",
-        "s:longitude",
-      ]);
-      location.seCorner.lat = parseFloatFromLocation([
-        "won:hasBoundingBox",
-        "won:hasSouthEastCorner",
-        "s:latitude",
-      ]);
-      location.seCorner.lng = parseFloatFromLocation([
-        "won:hasBoundingBox",
-        "won:hasSouthEastCorner",
-        "s:longitude",
-      ]);
-
-      if (
-        location.address &&
-        location.lat &&
-        location.lng &&
-        location.nwCorner.lat &&
-        location.nwCorner.lng &&
-        location.seCorner.lat &&
-        location.seCorner.lng
-      ) {
-        return Immutable.fromJS(location);
-      }
-
-      console.error(
-        "Cant parse location, data is an invalid location-object: ",
-        jsonldLocationImm.toJS()
-      );
-      return undefined;
-    },
-    generateHumanReadable: function({ value, includeLabel }) {
-      if (value) {
-        let humanReadable;
-        if (value.name) {
-          humanReadable = value.name;
-        } else if (value.address) {
-          humanReadable = value.address;
-        } else {
-          const locationLat = value.lat && value.lat.toFixed(6);
-          const locationLng = value.lng && value.lng.toFixed(6);
-          if (locationLat && locationLng) {
-            humanReadable = "@(" + locationLat + " , " + locationLng + ")";
-          }
-        }
-        if (humanReadable) {
-          return includeLabel
-            ? this.label + ": " + humanReadable.trim()
-            : humanReadable.trim();
-        }
-      }
-      return undefined;
-    },
-  },
-  person: {
-    identifier: "person",
-    label: "Person",
-    icon: "#ico36_detail_person",
-    placeholder: undefined,
-    component: "won-person-picker",
-    viewerComponent: "won-person-viewer",
-    messageEnabled: true,
-    parseToRDF: function({ value, identifier, contentUri }) {
-      if (!value) {
-        return { "foaf:person": undefined };
-      }
-      return {
-        "foaf:title": getIn(value, ["title"])
-          ? getIn(value, ["title"])
-          : undefined,
-        "foaf:name": getIn(value, ["name"])
-          ? getIn(value, ["name"])
-          : undefined,
-        "s:worksFor": getIn(value, ["company"])
-          ? {
-              "@id":
-                contentUri && identifier
-                  ? contentUri +
-                    "/" +
-                    identifier +
-                    "/" +
-                    generateIdString(10) +
-                    "/organization"
-                  : undefined,
-              "@type": "s:Organization",
-              "s:name": getIn(value, ["company"]),
-            }
-          : undefined,
-        "s:jobTitle": getIn(value, ["position"])
-          ? getIn(value, ["position"])
-          : undefined,
-      };
-    },
-    parseFromRDF: function(jsonLDImm) {
-      if (!jsonLDImm) return undefined;
-
-      let person = {
-        name: undefined,
-        title: undefined,
-        company: undefined,
-        position: undefined,
-        // bio: undefined,
-      };
-
-      person.name = won.parseFrom(jsonLDImm, ["foaf:name"], "xsd:string");
-      person.title = won.parseFrom(jsonLDImm, ["foaf:title"], "xsd:string");
-      person.company = won.parseFrom(
-        jsonLDImm,
-        ["s:worksFor", "s:name"],
-        "xsd:string"
-      );
-      person.position = won.parseFrom(jsonLDImm, ["s:jobTitle"], "xsd:string");
-
-      // if there's anything, use it
-      if (person.name || person.title || person.company || person.position) {
-        return Immutable.fromJS(person);
-      }
-      return undefined;
-    },
-    generateHumanReadable: function({ value, includeLabel }) {
-      const title = getIn(value, ["title"]);
-      const name = getIn(value, ["name"]);
-      const company = getIn(value, ["company"]);
-      const position = getIn(value, ["position"]);
-
-      let humanReadable;
-      if (title) {
-        humanReadable = title + " ";
-      }
-      if (name) {
-        if (humanReadable) {
-          humanReadable += name + " ";
-        } else {
-          humanReadable = name + " ";
-        }
-      }
-      if (company) {
-        if (humanReadable) {
-          humanReadable += "works at " + company + " ";
-        } else {
-          humanReadable = company + " ";
-        }
-      }
-      if (position) {
-        if (humanReadable) {
-          if (company) {
-            humanReadable += "as a " + position;
-          } else {
-            humanReadable += "is a " + position;
-          }
-        } else {
-          humanReadable = position;
-        }
-      }
-
-      if (humanReadable) {
-        return includeLabel
-          ? this.label + ": " + humanReadable.trim()
-          : humanReadable.trim();
-      }
-      return undefined;
-    },
-  },
-  travelAction: {
-    identifier: "travelAction",
-    label: "Route (From - To)",
-    icon: "#ico36_detail_travelaction",
-    placeholder: {
-      departure: "Start location",
-      destination: "Destination",
-    },
-    component: "won-travel-action-picker",
-    viewerComponent: "won-travel-action-viewer",
-    messageEnabled: true,
-    parseToRDF: function({ value, identifier, contentUri }) {
-      if (!value) {
-        return { "won:travelAction": undefined };
-      }
-
-      const randomTravelActionId = generateIdString(10);
-
-      return {
-        "won:travelAction": {
-          "@id":
-            contentUri && identifier
-              ? contentUri + "/" + identifier + "/" + randomTravelActionId
-              : undefined,
-          "@type": "s:TravelAction",
-          "s:fromLocation": !value.fromLocation
-            ? undefined
-            : {
-                "@id":
-                  contentUri && identifier
-                    ? contentUri +
-                      "/" +
-                      identifier +
-                      "/" +
-                      randomTravelActionId +
-                      "/fromLocation"
-                    : undefined,
-                "@type": "s:Place",
-                "s:geo": {
-                  "@id":
-                    contentUri && identifier
-                      ? contentUri +
-                        "/" +
-                        identifier +
-                        "/" +
-                        randomTravelActionId +
-                        "/fromLocation/geocoords"
-                      : undefined,
-                  "@type": "s:GeoCoordinates",
-                  "s:latitude": value.fromLocation.lat.toFixed(6),
-                  "s:longitude": value.fromLocation.lng.toFixed(6),
-                  "won:geoSpatial": {
-                    "@type":
-                      "http://www.bigdata.com/rdf/geospatial/literals/v1#lat-lon",
-                    "@value": `${value.fromLocation.lat.toFixed(
-                      6
-                    )}#${value.fromLocation.lng.toFixed(6)}`,
-                  },
-                },
-                "s:name": value.fromLocation.name,
-              },
-          "s:toLocation": !value.toLocation
-            ? undefined
-            : {
-                "@id":
-                  contentUri && identifier
-                    ? contentUri +
-                      "/" +
-                      identifier +
-                      "/" +
-                      randomTravelActionId +
-                      "/toLocation"
-                    : undefined,
-                "@type": "s:Place",
-                "s:geo": {
-                  "@id":
-                    contentUri && identifier
-                      ? contentUri +
-                        "/" +
-                        identifier +
-                        "/" +
-                        randomTravelActionId +
-                        "/toLocation/geocoords"
-                      : undefined,
-                  "@type": "s:GeoCoordinates",
-                  "s:latitude": value.toLocation.lat.toFixed(6),
-                  "s:longitude": value.toLocation.lng.toFixed(6),
-                  "won:geoSpatial": {
-                    "@type":
-                      "http://www.bigdata.com/rdf/geospatial/literals/v1#lat-lon",
-                    "@value": `${value.toLocation.lat.toFixed(
-                      6
-                    )}#${value.toLocation.lng.toFixed(6)}`,
-                  },
-                },
-                "s:name": value.toLocation.name,
-              },
-        },
-      };
-    },
-    parseFromRDF: function(jsonLDImm) {
-      const jsonTravelAction = jsonLDImm && jsonLDImm.get("won:travelAction");
-      if (!jsonTravelAction) return undefined;
-
-      const travelActionImm = Immutable.fromJS(jsonTravelAction);
-
-      let travelAction = {
-        fromAddress: undefined,
-        fromLocation: {
-          lat: undefined,
-          lng: undefined,
-        },
-        toAddress: undefined,
-        toLocation: {
-          lat: undefined,
-          lng: undefined,
-        },
-      };
-
-      travelAction.fromAddress = won.parseFrom(
-        travelActionImm,
-        ["s:fromLocation", "s:name"],
-        "xsd:string"
-      );
-
-      const parseFloatFromTravelAction = path =>
-        won.parseFrom(travelActionImm, path, "xsd:float");
-
-      travelAction.fromLocation.lat = parseFloatFromTravelAction([
-        "s:fromLocation",
-        "s:geo",
-        "s:latitude",
-      ]);
-      travelAction.fromLocation.lng = parseFloatFromTravelAction([
-        "s:fromLocation",
-        "s:geo",
-        "s:longitude",
-      ]);
-
-      travelAction.toAddress = won.parseFrom(
-        travelActionImm,
-        ["s:toLocation", "s:name"],
-        "xsd:string"
-      );
-
-      travelAction.toLocation.lat = parseFloatFromTravelAction([
-        "s:toLocation",
-        "s:geo",
-        "s:latitude",
-      ]);
-      travelAction.toLocation.lng = parseFloatFromTravelAction([
-        "s:toLocation",
-        "s:geo",
-        "s:longitude",
-      ]);
-
-      if (
-        (travelAction.fromAddress &&
-          travelAction.fromLocation.lat &&
-          travelAction.fromLocation.lng) ||
-        (travelAction.toAddress &&
-          travelAction.toLocation.lat &&
-          travelAction.toLocation.lng)
-      ) {
-        return Immutable.fromJS(travelAction);
-      } else {
-        console.error(
-          "Cant parse travelAction, data is an invalid travelAction-object: ",
-          travelActionImm.toJS()
-        );
-        return undefined;
-      }
-    },
-    generateHumanReadable: function({ value, includeLabel }) {
-      if (value && (value.fromLocation || value.toLocation)) {
-        const fromLocation = value.fromLocation;
-        const toLocation = value.toLocation;
-
-        let fromLocationName;
-        let toLocationName;
-
-        if (fromLocation && fromLocation.name) {
-          fromLocationName = fromLocation.name;
-        } else if (value && value.fromAddress) {
-          fromLocationName = value.fromAddress;
-        }
-
-        if (toLocation && toLocation.name) {
-          toLocationName = toLocation.name;
-        } else if (value && value.toAddress) {
-          toLocationName = value.toAddress;
-        }
-
-        let humanReadable;
-
-        if (fromLocationName) {
-          humanReadable = "from: " + fromLocationName + " ";
-        } else {
-          const fromLocationLat =
-            fromLocation && fromLocation.lat && fromLocation.lat.toFixed(6);
-          const fromLocationLng =
-            fromLocation && fromLocation.lng && fromLocation.lng.toFixed(6);
-          if (fromLocationLat && fromLocationLng) {
-            humanReadable =
-              "from: @(" + fromLocationLat + " , " + fromLocationLng + ") ";
-          }
-        }
-
-        if (toLocationName) {
-          if (humanReadable) {
-            humanReadable += "to: " + toLocationName + " ";
-          } else {
-            humanReadable = "to: " + toLocationName + " ";
-          }
-        } else {
-          const toLocationLat =
-            toLocation && toLocation.lat && toLocation.lat.toFixed(6);
-          const toLocationLng =
-            toLocation && toLocation.lng && toLocation.lng.toFixed(6);
-          if (toLocationLat && toLocationLng) {
-            if (humanReadable) {
-              humanReadable +=
-                "to: @(" + toLocationLat + " , " + toLocationLng + ") ";
-            } else {
-              humanReadable +=
-                "to: @(" + toLocationLat + " , " + toLocationLng + ") ";
-            }
-          }
-        }
-
-        if (humanReadable) {
-          return includeLabel
-            ? this.label + ": " + humanReadable.trim()
-            : humanReadable.trim();
-        }
       }
       return undefined;
     },
@@ -979,403 +226,27 @@ export const details = {
       return undefined;
     },
   },
-  files: {
-    identifier: "files",
-    label: "Files",
-    icon: "#ico36_detail_files",
-    placeholder: "",
-    accepts: "",
-    multiSelect: true,
-    component: "won-file-picker",
-    viewerComponent: "won-file-viewer",
-    messageEnabled: true,
-    parseToRDF: function({ value, identifier, contentUri }) {
-      if (!value) {
-        return { "won:hasFile": undefined };
-      }
-      let payload = [];
-      value.forEach(file => {
-        //TODO: SAVE CORRECT RDF THIS METHOD
-        if (file.name && file.type && file.data) {
-          let f = {
-            "@id":
-              contentUri && identifier
-                ? contentUri + "/" + identifier + "/" + generateIdString(10)
-                : undefined,
-            "@type": "s:FileObject",
-            "s:name": file.name,
-            "s:type": file.type,
-            "s:data": file.data,
-          };
+};
 
-          payload.push(f);
-        }
-      });
-      if (payload.length > 0) {
-        return { "won:hasFile": payload };
-      }
-      return { "won:hasFile": undefined };
-    },
-    parseFromRDF: function(jsonLDImm) {
-      const files = jsonLDImm && jsonLDImm.get("won:hasFile");
-      let parsedFiles = [];
+export const details = {
+  title: basicDetails.title,
+  description: basicDetails.description,
+  tags: basicDetails.tags,
 
-      if (Immutable.List.isList(files)) {
-        files &&
-          files.forEach(file => {
-            //TODO: RETRIEVE FROM CORRECT RDF THIS METHOD
-            let f = {
-              name: get(file, "s:name"),
-              type: get(file, "s:type"),
-              data: get(file, "s:data"),
-            };
-            if (f.name && f.type && f.data) {
-              parsedFiles.push(f);
-            }
-          });
-      } else {
-        let f = {
-          name: get(files, "s:name"),
-          type: get(files, "s:type"),
-          data: get(files, "s:data"),
-        };
-        if (f.name && f.type && f.data) {
-          return Immutable.fromJS([f]);
-        }
-      }
-      if (parsedFiles.length > 0) {
-        return Immutable.fromJS(parsedFiles);
-      }
-      return undefined;
-    },
-    generateHumanReadable: function({ value, includeLabel }) {
-      if (value) {
-        let humanReadable = "";
-        if (value.length > 1) {
-          humanReadable = value.length + " Files";
-        } else {
-          humanReadable = value[0].name;
-        }
+  fromDatetime: timeDetails.fromDatetime,
+  throughDatetime: timeDetails.throughDatetime,
+  datetimeRange: timeDetails.datetimeRange,
 
-        return includeLabel ? this.label + ": " + humanReadable : humanReadable;
-      }
-      return undefined;
-    },
-  },
-  images: {
-    identifier: "images",
-    label: "Images",
-    icon: "#ico36_detail_media",
-    placeholder: "",
-    accepts: "image/*",
-    multiSelect: true,
-    component: "won-file-picker",
-    viewerComponent: "won-file-viewer",
-    messageEnabled: true,
-    parseToRDF: function({ value, identifier, contentUri }) {
-      if (!value) {
-        return { "won:hasImage": undefined };
-      }
-      let payload = [];
-      value.forEach(image => {
-        //TODO: SAVE CORRECT RDF THIS METHOD
-        if (image.name && image.type && image.data) {
-          let img = {
-            "@id":
-              contentUri && identifier
-                ? contentUri + "/" + identifier + "/" + generateIdString(10)
-                : undefined,
-            "@type": "s:ImageObject",
-            "s:name": image.name,
-            "s:type": image.type,
-            "s:data": image.data,
-          };
+  location: locationDetails.location,
+  travelAction: locationDetails.travelAction,
 
-          payload.push(img);
-        }
-      });
-      if (payload.length > 0) {
-        return { "won:hasImage": payload };
-      }
-      return { "won:hasImage": undefined };
-    },
-    parseFromRDF: function(jsonLDImm) {
-      const images = jsonLDImm && jsonLDImm.get("won:hasImage");
-      let imgs = [];
+  person: personDetails.person,
 
-      if (Immutable.List.isList(images)) {
-        images &&
-          images.forEach(image => {
-            //TODO: RETRIEVE FROM CORRECT RDF THIS METHOD
-            let img = {
-              name: get(image, "s:name"),
-              type: get(image, "s:type"),
-              data: get(image, "s:data"),
-            };
-            if (img.name && img.type && img.data) {
-              imgs.push(img);
-            }
-          });
-      } else {
-        let img = {
-          name: get(images, "s:name"),
-          type: get(images, "s:type"),
-          data: get(images, "s:data"),
-        };
-        if (img.name && img.type && img.data) {
-          return Immutable.fromJS([img]);
-        }
-      }
-      if (imgs.length > 0) {
-        return Immutable.fromJS(imgs);
-      }
-      return undefined;
-    },
-    generateHumanReadable: function({ value, includeLabel }) {
-      if (value) {
-        let humanReadable = "";
-        if (value.length > 1) {
-          humanReadable = value.length + " Images";
-        } else {
-          humanReadable = value[0].name;
-        }
+  files: fileDetails.files,
+  images: fileDetails.images,
+  bpmnWorkflow: fileDetails.bpmnWorkflow,
+  petrinetWorkflow: fileDetails.petrinetWorkflow,
 
-        return includeLabel ? this.label + ": " + humanReadable : humanReadable;
-      }
-      return undefined;
-    },
-  },
-  bpmnWorkflow: {
-    identifier: "bpmnWorkflow",
-    label: "BPMN",
-    icon: "#ico36_detail_workflow",
-    placeholder: "",
-    //accepts: "application/octet-stream",
-    accepts: "",
-    component: "won-workflow-picker",
-    viewerComponent: "won-workflow-viewer",
-    messageEnabled: true,
-    parseToRDF: function({ value, identifier, contentUri }) {
-      if (value && value.name && value.data) {
-        //do not check for value.type might not be present on some systems
-        let workflow = {
-          "@id":
-            contentUri && identifier
-              ? contentUri + "/" + identifier + "/" + generateIdString(10)
-              : undefined,
-          "@type": "s:FileObject",
-          "s:name": value.name,
-          "s:type": value.type,
-          "s:data": value.data,
-        };
-
-        return { "won:hasBpmnWorkflow": workflow };
-      }
-      return { "won:hasBpmnWorkflow": undefined };
-    },
-    parseFromRDF: function(jsonLDImm) {
-      const wflw = jsonLDImm && jsonLDImm.get("won:hasBpmnWorkflow");
-
-      let workflow = {
-        name: get(wflw, "s:name"),
-        type: get(wflw, "s:type"),
-        data: get(wflw, "s:data"),
-      };
-      if (workflow.name && workflow.data) {
-        //do not check for value.type might not be present on some systems
-        return Immutable.fromJS(workflow);
-      }
-
-      return undefined;
-    },
-    generateHumanReadable: function({ value, includeLabel }) {
-      if (value && value.name) {
-        return includeLabel ? this.label + ": " + value.name : value.name;
-      }
-      return undefined;
-    },
-  },
-  petrinetWorkflow: {
-    identifier: "petrinetWorkflow",
-    label: "Petrinet",
-    icon: "#ico36_detail_workflow",
-    placeholder: "",
-    //accepts: "application/octet-stream",
-    accepts: "",
-    component: "won-petrinet-picker",
-    viewerComponent: "won-petrinet-viewer",
-    messageEnabled: true,
-    parseToRDF: function({ value, identifier, contentUri }) {
-      if (value && value.name && value.data) {
-        //do not check for value.type might not be present on some systems
-        let workflow = {
-          "@id":
-            contentUri && identifier
-              ? contentUri + "/" + identifier + "/" + generateIdString(10)
-              : undefined,
-          "@type": "s:FileObject",
-          "s:name": value.name,
-          "s:type": value.type,
-          "s:data": value.data,
-        };
-
-        return { "won:hasPetrinet": workflow };
-      }
-      return { "won:hasPetrinet": undefined };
-    },
-    parseFromRDF: function(jsonLDImm) {
-      const wflw = jsonLDImm && jsonLDImm.get("won:hasPetrinet");
-
-      let workflow = {
-        name: get(wflw, "s:name"),
-        type: get(wflw, "s:type"),
-        data: get(wflw, "s:data"),
-      };
-      if (workflow.name && workflow.data) {
-        //do not check for value.type might not be present on some systems
-        return Immutable.fromJS(workflow);
-      }
-
-      return undefined;
-    },
-    generateHumanReadable: function({ value, includeLabel }) {
-      if (value && value.name) {
-        return includeLabel ? this.label + ": " + value.name : value.name;
-      }
-      return undefined;
-    },
-  },
-  pricerange: {
-    identifier: "pricerange",
-    label: "Price range",
-    minLabel: "Min",
-    maxLabel: "Max",
-    minPlaceholder: "Min Price",
-    maxPlaceholder: "Max Price",
-    icon: "#ico36_detail_price",
-    currency: [
-      { value: "EUR", label: "€", default: true },
-      { value: "USD", label: "$" },
-      { value: "GBP", label: "£" },
-    ],
-    unitCode: [
-      { value: "MON", label: "per month" },
-      { value: "WEE", label: "per week" },
-      { value: "DAY", label: "per day" },
-      { value: "HUR", label: "per hour" },
-      { value: "", label: "total", default: true },
-    ],
-    component: "won-price-range-picker",
-    viewerComponent: "won-price-viewer",
-    messageEnabled: true,
-    parseToRDF: function({ value, identifier, contentUri }) {
-      if (!value || !(value.min || value.max) || !value.currency) {
-        return { "s:priceSpecification": undefined };
-      }
-      return {
-        "s:priceSpecification": {
-          "@id":
-            contentUri && identifier
-              ? contentUri + "/" + identifier + "/" + generateIdString(10)
-              : undefined,
-          "@type": "s:CompoundPriceSpecification",
-          "s:minPrice": value.min && [
-            { "@value": value.min, "@type": "xsd:float" },
-          ],
-          "s:maxPrice": value.max && [
-            { "@value": value.max, "@type": "xsd:float" },
-          ],
-          "s:priceCurrency": value.currency,
-          "s:unitCode": value.unitCode,
-          "s:description": "total rent per month in between min/max",
-        },
-      };
-    },
-    parseFromRDF: function(jsonLDImm) {
-      const minRent = won.parseFrom(
-        jsonLDImm,
-        ["s:priceSpecification", "s:minPrice"],
-        "xsd:float"
-      );
-      const maxRent = won.parseFrom(
-        jsonLDImm,
-        ["s:priceSpecification", "s:maxPrice"],
-        "xsd:float"
-      );
-      const currency = won.parseFrom(
-        jsonLDImm,
-        ["s:priceSpecification", "s:priceCurrency"],
-        "xsd:string"
-      );
-      const unitCode = won.parseFrom(
-        jsonLDImm,
-        ["s:priceSpecification", "s:unitCode"],
-        "xsd:string"
-      );
-
-      if (!minRent && !maxRent) {
-        return undefined;
-      } else if (!currency) {
-        return undefined;
-      } else {
-        // if there's anything, use it
-        return Immutable.fromJS({
-          min: minRent,
-          max: maxRent,
-          currency: currency,
-          unitCode: unitCode,
-        });
-      }
-    },
-    generateHumanReadable: function({ value, includeLabel }) {
-      if (value) {
-        const min = value.min;
-        const max = value.max;
-
-        let amount;
-        if (min && max) {
-          amount = min + " - " + max;
-        } else if (min) {
-          amount = "at least " + min;
-        } else if (max) {
-          amount = "at most " + max;
-        } else {
-          return undefined;
-        }
-
-        let currencyLabel = undefined;
-        let unitCodeLabel = undefined;
-
-        this.currency &&
-          this.currency.forEach(curr => {
-            if (curr.value === value.currency) {
-              currencyLabel = curr.label;
-            }
-          });
-        currencyLabel = currencyLabel || value.currency;
-
-        this.unitCode &&
-          this.unitCode.forEach(uc => {
-            if (uc.value === value.unitCode) {
-              unitCodeLabel = uc.label;
-            }
-          });
-        unitCodeLabel = unitCodeLabel || value.unitCode;
-
-        if (unitCodeLabel) {
-          return (
-            (includeLabel ? this.label + ": " + amount : amount) +
-            currencyLabel +
-            " " +
-            unitCodeLabel
-          );
-        } else {
-          return (
-            (includeLabel ? this.label + ": " + amount : amount) + currencyLabel
-          );
-        }
-      }
-      return undefined;
-    },
-  },
+  pricerange: priceDetails.pricerange,
+  price: priceDetails.price,
 };
